mcp/requirements.txt
mcp[cli]>=1.10,<2
fastapi>=0.115
uvicorn[standard]>=0.32
supabase>=2.7
openai>=1.50
pydantic>=2.9
pydantic-settings>=2.6
python-dotenv>=1.0
sentry-sdk[fastapi]>=2.0
pyjwt>=2.9
httpx>=0.27
